0) Pygame. Done
game main loop
