+++ roadmap
@@ -2,11 +2,18 @@
         -ROADMAP-
 
 0) Pygame:
-    pygame settings and interaction variables
+    pygame settings
     pygame main loop  (drawing and interaction)
 
-1) Display dev info:
-    toggle: info, screen axis, note, mouse pos
+
+1) Dev info and qol features:
+    text writing
+
+    toggle info
+    toggle screen axis, note, mouse pos
+
+    change volume
+    toggle mute
 
 
 2) Music variables:
@@ -17,45 +24,64 @@
     Map note: value
 
 
-3) display piano keys:
+3) Sound files
+    note sound
+
+
+4) Sound playing:
+    play note
+
+
+5) Test all sound files
+
+
+6) display piano keys:
     drawing surface
     draw keys
 
 
-4) Class Button:
+7) Class Button:
     .draw()
     .write() if text button
     .ispressed()
 
 
-5) User input variable:
+8) User input variable:
     input note
     input chord
 
 
-6) Detect mouse input:
+8) Detect mouse input:
     get pressed key
     get mouse input
     store input note
     display input note
 
 
-7) Display chord buttons:
+9) Sound playing:
+    press space to play selected note
+
+
+10) Display chord buttons:
     drawing surface
-    draw buttons
+    draw buttons  (using grid for placement)
 
 
-8) Detect mouse input:
+11) Detect mouse input:
     get pressed chord
     get mouse input
     store input chord
     display input chord
 
 
-9) Highlight pressed keys
-10) Highlight pressed chord
+12) Highlight pressed keys
+13) Highlight pressed chord
 
-11) Bonus features:
+14) Sound playing:
+    play chords
+
+
+ Bonus features:
     display note on music staff
 
 
@@ -64,6 +90,13 @@
 
 settings and interaction variables:
     screen width/height, Clock
+
+drawing variables:
+
+    color list
+    toolbar spec
+
+
 
 music variables:
