+++ src/drawing_variables.py
@@ -1,0 +1,27 @@
+import pygame.font
+
+from .settings import WIDTH, HEIGHT
+
+
+
+colors = {
+    'light_grey': '#788080',
+    'grey': '#606060',
+    'green': '#405700',
+    'dark_green': '#1B4510',
+    'warm_yellow': '#806010',
+    'green1': '#007030',
+    'light_blue1': '#606885',
+    'light_blue2': '#607579',
+    'brown': '#604000'
+}
+
+bg_color = '#00222A'
+
+
+FONT15 = pygame.font.SysFont('arial', 15)
+FONT20 = pygame.font.SysFont('arial', 20)
+FONT25 = pygame.font.SysFont('arial', 25)
+
+
+
