+++ src/interactive_variables.py
@@ -1,0 +1,35 @@
+from src.settings import WIDTH, HEIGHT
+
+
+# mouse position:
+pos = pos_x, pos_y = round(WIDTH/2), round(HEIGHT/2)
+
+
+# create piano keys
+# create chord buttons
+
+
+# init input variable:
+#   chosen note
+#   selected key
+#   selected chord
+
+
+# settings:
+#   BPM
+#   volume
+
+VOLUME = 0.3
+
+
+
+# dev controls:
+
+TOGGLE_INFO = False
+
+TOGGLE_MUTE = False
+TOGGLE_AXIS = False
+TOGGLE_CHOSEN_NOTE = False
+TOGGLE_FONT = False
+TOGGLE_FLAT = False
+TOGGLE_POS = True
