+++ src/settings.py
@@ -1,0 +1,17 @@
+import pygame
+
+pygame.init()
+
+
+WIDTH, HEIGHT = 700,700
+
+WIN = pygame.display.set_mode((WIDTH, HEIGHT))
+
+
+clock = pygame.time.Clock()
+FPS = 15
+
+
+BPM = 120
+
+
