--- utils/__init__.py
@@ -1,15 +1,0 @@
-from .settings import *
-
-from .drawing_variables import *
-
-from .sound_playing import *
-
-from.music_functions import get_allchords
-
-from.music_variables import *
-
-from.interactive_variables import *
-
-
-pygame.font.init()
-
