--- utils/settings.py
@@ -1,15 +1,0 @@
-from .music_variables import notes_name, value_mapping, Notes_spec
-
-import pygame
-
-from .formatting import pretty_print_note
-
-
-
-chosen_note = None
-chord = []
-
-BPM = 120
-VOLUME = 0.3
-
-
